refactor: reorganize root scripts and update references
