+++ .gitignore
@@ -20,6 +20,11 @@ archive/*.tar.gz
 *.tmp
 # Test artifacts
 tmp/
+# Build and report artifacts
+builds/
+results/
+reports/
+artifacts/
 # Git LFS temporary files
 *.lfs.tmp
 *.lfs.lock
