+++ AGENTS.md
@@ -59,6 +59,7 @@ When using the terminal or editing files, the agent must adhere to the following
   * These output management strategies are **mandatory** for high-risk commands. Always anticipate if a command could produce very long lines or a flood of data. When in doubt, err on the side of caution and apply one of the above methods (or use filtering commands like `grep`, `cut`, or `awk` to reduce output size). **Never** allow an unbounded output stream to print directly to the terminal.
   * **Self-Healing Recovery**: If the terminal shows `Error: Output for session 'shell' contained a line exceeding the max of 1600 bytes`, immediately start a new session, run the standard setup steps again (reactivate the virtual environment and ensure `/usr/local/bin/clw` exists), then repeat the previous command with a safe output method such as piping through `clw` or logging and chunking.
 * **Anti-Recursion & Backups**: Never write backup files or any output to the repository workspace itself. All backup or log files (if any) should go to the designated external backup directory (see **GH\_COPILOT\_BACKUP\_ROOT** above). The agent should validate paths (using any provided `validate_enterprise_operation()` or similar, if available) before performing file writes, to ensure compliance with anti-recursion rules.
+* **Artifact Management**: Runtime directories such as `builds/`, `logs/`, `results/`, `reports/`, and `tmp/` must not be committed. Store temporary outputs under `artifacts/` or archive them with `artifact_manager.py` when needed.
 * **Dual Execution Pattern**: When appropriate, follow the “Dual Copilot” approach – i.e., for critical operations, simulate having a primary action followed by a validation step. (For example, after generating a script, perform a dry-run or static analysis to verify it before finalizing.)
 * **Adhere to Protocol**: Always incorporate required safety checks (e.g., zero-byte file checks, recursion detection, etc.) as described in project protocols. Before concluding any major operation, ensure you run any integrity validations mandated by the project’s standards.
