+++ archive/legacy_root_scripts/quantum/quantum_algorithms_functional.py
