+++ archive/legacy_root_scripts/quantum/quantum_neural_networks_predictive_maintenance.py
