+++ archive/legacy_root_scripts/quantum/quantum_optimizer.py
