+++ ml_pattern_recognition/__init__.py
@@ -1,7 +1,5 @@
 """ML pattern recognition package."""
 
-from .recognizer import PatternRecognizer, SklearnPatternModel
-
 from .recognizer import PatternRecognizer, SklearnPatternModel, train_pipeline
 
 __all__ = ["PatternRecognizer", "SklearnPatternModel", "train_pipeline"]
