+++ scripts/validation/performance_validation_framework.py
@@ -17,7 +17,7 @@ from pathlib import Path
 
 from tqdm import tqdm
 
-from database_driven_flake8_corrector_functional import DatabaseDrivenFlake8CorrectorFunctional
+from scripts.database.database_driven_flake8_corrector_functional import DatabaseDrivenFlake8CorrectorFunctional
 from scripts.utilities.quantum_algorithms_functional import (
     run_grover_search,
     run_kmeans_clustering,
