+++ template_engine/db_first_code_generator.py
@@ -35,7 +35,7 @@ from .learning_templates import get_lesson_templates
 
 # Quantum scoring helper
 try:
-    from quantum_algorithm_library_expansion import (
+    from quantum.quantum_algorithm_library_expansion import (
         quantum_text_score,
         quantum_similarity_score,
         quantum_cluster_score,
