+++ tests/test_session_management_consolidation_executor.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from session_management_consolidation_executor import EnterpriseUtility
+from scripts.session.session_management_consolidation_executor import EnterpriseUtility
 
 
 def test_consolidation_executor_pass(tmp_path, monkeypatch):
