+++ tests/test_session_protocol_validator.py
@@ -3,7 +3,7 @@
 from scripts.utilities.unified_session_management_system import (
     UnifiedSessionManagementSystem,
 )
-from session_protocol_validator import SessionProtocolValidator
+from scripts.session.session_protocol_validator import SessionProtocolValidator
 
 
 def test_startup_detects_zero_byte(tmp_path, monkeypatch):
