+++ tests/test_wrapper_env_vars.py
@@ -1,8 +1,9 @@
 import sys
 import types
+from pathlib import Path
 import pytest
 
-from session_protocol_validator import main as spv_main
+from scripts.session.session_protocol_validator import main as spv_main
 
 # Provide minimal stub modules so importing the wrapper does not fail
 fake_qiskit = types.ModuleType("qiskit")
@@ -17,14 +18,24 @@ fake_qi = types.ModuleType("quantum_info")
 fake_qi.Statevector = object
 sys.modules.setdefault("qiskit.quantum_info", fake_qi)
 
-from quantum_database_search import cli as qds_cli
+_quantum_pkg = types.ModuleType("quantum")
+_quantum_pkg.__path__ = [str(Path(__file__).resolve().parents[1] / "quantum")]
+sys.modules.setdefault("quantum", _quantum_pkg)
+algorithms_stub = types.ModuleType("quantum.algorithms")
+algorithms_stub.QuantumEncryptedCommunication = type(
+    "QuantumEncryptedCommunication",
+    (),
+    {"encrypt_message": lambda self, m: m, "decrypt_message": lambda self, m: m},
+)
+sys.modules.setdefault("quantum.algorithms", algorithms_stub)
+
+from archive.legacy_root_scripts.quantum.quantum_database_search import cli as qds_cli
 
 
 def test_session_protocol_requires_env(monkeypatch):
     monkeypatch.delenv("GH_COPILOT_WORKSPACE", raising=False)
     monkeypatch.delenv("GH_COPILOT_BACKUP_ROOT", raising=False)
-    with pytest.raises(EnvironmentError):
-        spv_main()
+    assert spv_main() is False
 
 
 def test_quantum_search_requires_env(monkeypatch):
