--- tmp/test_artifacts.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:1aecd5a2256def2f57b3b67be0fb7cd43383e7b40d9f5c95763db59edda67d66
-size 1108941
